fix: moved AddLinkToDatabaseResponse.java, GenericResponse.java, ListLinksResponse.java, RemoveLinkFromDatabaseResponse.java from models to api.dto.response package
