+++ bot/src/main/java/ru/bytebosses/bot/api/dto/response/ListLinksResponse.java
@@ -1,0 +1,6 @@
+package ru.bytebosses.bot.api.dto.response;
+
+import ru.bytebosses.bot.models.Link;
+import java.util.List;
+
+public record ListLinksResponse(List<Link> links) {}
