+++ bot/src/main/java/ru/bytebosses/bot/api/dto/response/RemoveLinkFromDatabaseResponse.java
@@ -1,4 +1,4 @@
-package ru.bytebosses.bot.models;
+package ru.bytebosses.bot.api.dto.response;
 
 import java.net.URI;
